// src/types/common/CpDatType.ts
interface CpData {
    cP: string;
    eX: string;
    pL: string;
    eurRub: string;
    customs: string;
    round: string;
    marg: string | number;
    salesWb: number;
    salesYa: number;
    salesOz: number;
    salesMm: number;
    salesAv: number;
    fixedExp: number;
    quart: number;
    priceIndex: string;
    middlePercent: string;
    middlePercent2: string;
    minPer: string;
    lS: string;
  }

  export default CpData;